RISC-V: KVM: Add a helper function to check if TVM vcpu is blocked
The TSM may move a vcpu to blocking state if it requires any fence to
complete any pending transaction. The TSM notifies the host about that
via the return value of SBI_EXT_COVH_TVM_VCPU_RUN call. The host is
expected to issue a tvm fence operations before resuming the vcpu.
Otherwise, the vcpu run would fail for that vcpu.

Signed-off-by: Rajnesh Kanwal <rkanwal@rivosinc.com>
Signed-off-by: Atish Patra <atishp@rivosinc.com>