cpufreq: CPPC: add support for fully autonomous mode
ACPI spec [1] defines a mode (section 8.4.7.1.5) were the platform
controls the performance states autonomously. Add support for this mode
which uses the policy min/max as well as an energy performance
preference.

The energy performance preference is defined as a number betwenn 0x0 and
0xFF. This implementation provides 3 predefined values to match the
amd_pstate and intel_pstate driver. This values are "performance",
"balanced" and "powersave".

[1] https://uefi.org/htmlspecs/ACPI_Spec_6_4_html/08_Processor_Configuration_and_Control/declaring-processors.html

Signed-off-by: Clément Léger <cleger@rivosinc.com>