riscv: add support for SBI Supervisor Software Events extension
The SBI SSE extension allows the supervisor software to be notified by
the SBI of specific events that are not maskable. The context switch is
handled entirely by the firmware which will save all registers into a
supervisor provided context and restore a set of registers from this
same context. This context contains all the RV GP registers and thus
will also switch the stack.

Since each events (and each CPU for local events) have their own
context (and thus their own set of register), allocate a stack for each
event (and for each cpu for local events). This can be modified if
needed to use a single stack per cpu for SSE event but will require a
bit more additional plumbing.

When completing the event, if we were coming from kernel with interrupts
disabled, simply return there. If coming from userspace or kernel with
interrupts enabled, simulate an interrupt exception to allow delivery of
signal to user task. For instance this can happen, when a RAS event has
been generated and a SIGBUS has been sent to a task.

Signed-off-by: Clément Léger <cleger@rivosinc.com>